refactor : mapper util로 이동
